# tasks/backend/GO/gremiha3/docs/swagger.yaml
definitions:
...
  httpserver.ProductRequest:
    properties:
      name:
        example: синхрофазотрон
        type: string
      price:
        example: 1245.65
        type: number
      provider_id:
        example: 1
        type: integer
      stock:
        example: 435
        type: integer
    type: object
  httpserver.ProductResponse:
    properties:
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      provider_id:
        type: integer
      stock:
        type: integer
    type: object
  httpserver.ProviderRequest:
    properties:
      name:
        example: Microsoft
        type: string
      origin:
        example: Vietnam
        type: string
    type: object
  httpserver.ProviderResponse:
    properties:
      id:
        type: integer
      name:
        type: string
      origin:
        type: string
    type: object
...
  httpserver.UserResponse:
    properties:
      id:
        type: integer
      login:
        type: string
      password:
        type: string
      role:
        type: string
      token:
        type: string
    type: object
info:
  contact: {}
  description: An Shop service API in Go using Gin framework
  title: Shop Service API
  version: "1.0"
paths:
...
  /product:
    get:
      description: Return product with "id" number.
      parameters:
      - default: "1"
        description: id of the product
        example: "1"
        in: query
        name: id
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpserver.ProductResponse'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Посмотреть товар по его id.
      tags:
      - Product
    post:
      description: Save register data of user in Repo.
      parameters:
      - description: Create product
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/httpserver.ProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpserver.ProductResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Добавить товар.
      tags:
      - Product
  /products:
    get:
      description: Return products list.
      parameters:
      - default: "10"
        description: limit records on page
        example: "10"
        in: query
        name: limit
        required: true
        type: string
      - default: "0"
        description: start of record output
        example: "0"
        in: query
        name: offset
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/httpserver.ProductResponse'
            type: array
        "404":
          description: Not Found
          schema:
            type: string
      summary: Получить список всех товаров.
      tags:
      - Product
  /provider:
    get:
      description: Return Provider with "id" number.
      parameters:
      - description: id of the provider
        example: "1"
        in: query
        name: id
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpserver.ProviderResponse'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Посмотреть постащика по его id.
      tags:
      - Provider
    post:
      description: Save register data of user in Repo.
      parameters:
      - description: Create Provider
        in: body
        name: Provider
        required: true
        schema:
          $ref: '#/definitions/httpserver.ProviderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpserver.ProviderResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Добавить поставщика.
      tags:
      - Provider
  /providers:
    get:
      description: Return Providers list.
      parameters:
      - default: "10"
        description: limit records on page
        example: "10"
        in: query
        name: limit
        required: true
        type: string
      - default: "0"
        description: start of record output
        example: "0"
        in: query
        name: offset
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/httpserver.ProviderResponse'
            type: array
        "404":
          description: Not Found
          schema:
            type: string
      summary: Получить список всех поставщиков.
      tags:
      - Provider
...
  /user:
    get:
      description: Получить пользователя по его id ли логину.
      parameters:
      - default: "1"
        description: id of the user
        example: "1"
        in: query
        name: id
        type: string
      - default: cmd@cmd.ru
        description: login of the user
        example: cmd@cmd.ru
        in: query
        name: login
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpserver.UserResponse'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Посмотреть пользователя по его id или логину.
      tags:
      - User
  /users:
    get:
      description: Return users list.
      parameters:
      - default: "10"
        description: limit records on page
        example: "10"
        in: query
        name: limit
        required: true
        type: string
      - default: "0"
        description: start of record output
        example: "0"
        in: query
        name: offset
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/httpserver.UserResponse'
            type: array
        "404":
          description: Not Found
          schema:
            type: string
      summary: Получить список всех пользователей.
      tags:
      - User
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"